ft(order): Generate order db models

+++ db/migration/000001_init_schema.up.sql
@@ -18,7 +18,7 @@ CREATE TABLE books (
 
 CREATE TABLE orders (
   id INTEGER PRIMARY KEY,
-  price integer NOT NULL,
+  price REAL NOT NULL,
   book_id integer NOT NULL,
   user_id integer NOT NULL,
   status varchar NOT NULL DEFAULT 'PENDING',
